move resolver stuff to its own package; update transportPool so it closes its resolver when shutting down
